Refactor admin side
