+++ app/controllers/decidim/polis/admin/application_controller.rb
@@ -8,7 +8,7 @@ module Decidim
       #
       # Note that it inherits from `Decidim::Admin::Components::BaseController`, which
       # override its layout and provide all kinds of useful methods.
-      class ApplicationController < Decidim::Admin::Components::BaseController
+      class ApplicationController < Decidim::Admin::ApplicationController
       end
     end
   end
