--- app/views/decidim/admin/polis_config/edit.html.erb
@@ -1,7 +1,0 @@
-<%= decidim_form_for(@form, html: { class: "form edit_polis_config" }, url: organization_polis_config_update_path, method: :put) do |f| %>
-  <%= render partial: 'form', object: f %>
-  <div class="button--double form-general-submit">
-    <%= f.submit t(".update") %>
-  </div>
-<% end %>
-
