+++ app/views/decidim/polis/admin/configuration/edit.html.erb
@@ -1,0 +1,7 @@
+<%= decidim_form_for(@form, html: { class: "form edit_polis_config" }, url: decidim_admin_polis.organization_configuration_update_path, method: :put) do |f| %>
+  <%= render partial: 'form', object: f %>
+  <div class="button--double form-general-submit">
+    <%= f.submit t(".update") %>
+  </div>
+<% end %>
+
