+++ app/views/layouts/decidim/admin/settings.html.erb
@@ -28,8 +28,8 @@
       <li <% if allowed_to?(:update, :help_sections) && is_active_link?(decidim_admin.help_sections_path) %> class="is-active" <% end %>>
         <%= link_to t("menu.help_sections", scope: "decidim.admin"), decidim_admin.help_sections_path %>
       </li>
-      <li <% if is_active_link?(organization_polis_config_edit_path) %> class="is-active" <% end %>>
-        <%= link_to "Pol.is", organization_polis_config_edit_path %>
+      <li <% if is_active_link?(decidim_admin_polis.organization_configuration_edit_path) %> class="is-active" <% end %>>
+        <%= link_to "Pol.is", decidim_admin_polis.organization_configuration_edit_path %>
       </li>
     </ul>
   </div>
