+++ db/migrate/20180301065746_add_polis_site_id_to_organizations.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class AddPolisSiteIdToOrganizations < ActiveRecord::Migration[5.1]
   def change
     add_column :decidim_organizations, :polis_site_id, :string
