+++ lib/decidim/polis/test/factories.rb
@@ -5,7 +5,7 @@ require "decidim/core/test/factories"
 FactoryBot.define do
   factory :polis_component, parent: :component do
     name { Decidim::Components::Namer.new(participatory_space.organization.available_locales, :polis).i18n_name }
-    manifest_name :polis
+    manifest_name { :polis }
     participatory_space { create(:participatory_process, :with_steps) }
   end
